Hydra::Collections::SearchBehaviors module
Allows you to include these behaviors within the search builder of your
own application.
